Создал конструкторы и подправил классы

+++ src/main/java/ru/ruru/education/Animal.java
@@ -3,10 +3,9 @@ package ru.ruru.education;
 class Animal {
 
     String name; // имя
-    byte age;     // возраст
+    int age;     // возраст
     byte height; // высота в холке в см
     byte lnTail; // длина хвоста
     boolean playful; // игривость
-
 
 }
